% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19SeverityPrediction1.R
\docType{package}
\name{Covid19SeverityPrediction1}
\alias{Covid19SeverityPrediction1}
\alias{Covid19SeverityPrediction1-package}
\title{Covid19SeverityPrediction1}
\description{
Covid19SeverityPrediction1
}
